+++ lib/lightning/workflows.ex
@@ -334,20 +334,26 @@ defmodule Lightning.Workflows do
       %Ecto.Changeset{data: %Workflow{}}
 
   """
-  def mark_for_deletion(workflow, _attrs \\ %{}) do
+  def mark_for_deletion(workflow, actor, _attrs \\ %{}) do
     workflow_triggers_query =
       from(t in Lightning.Workflows.Trigger,
         where: t.workflow_id == ^workflow.id
       )
 
-    Repo.transaction(fn ->
+    Multi.new()
+    |> Multi.update(
+      :workflow,
       Workflow.request_deletion_changeset(workflow, %{
         "deleted_at" => DateTime.utc_now()
       })
-      |> Repo.update()
-
-      Repo.update_all(workflow_triggers_query, set: [enabled: false])
-    end)
+    )
+    |> Multi.insert(:audit, Audit.marked_for_deletion(workflow.id, actor))
+    |> Multi.update_all(
+      :disable_triggers,
+      workflow_triggers_query,
+      set: [enabled: false]
+    )
+    |> Repo.transaction()
     |> tap(fn result ->
       with {:ok, _} <- result do
         workflow